Writing Serialization for Map Works
Implement a reader for Local and remote and Implement the Writer to remote. for Editor Time at the moment.

Work on a Chunk Layer this layer would work side by side in the tile map and it would be 50x50 and then when it comes to serialization I can just pull the data.

Chunk Layer would required to be underneath a TileLayer Strict